refactor(frontend-reader): move chapter layout above [chapter] segment for partial rendering
layout.tsx が [chapter] dynamic segment の配下にあると、チャプター遷移時に
layout も再実行されて TOC が再レンダリングされる。layout を (detail)/ route
group 経由で [chapter] の上位に配置することで、Next.js の partial rendering
が効き TOC の DOM と fetch が保持される。
