+++ applications/frontend/admin/src/app/series/[slug]/chapters/(detail)/[chapter]/page.tsx
