+++ applications/frontend/reader/src/app/series/[slug]/chapters/(detail)/[chapter]/page.tsx
