+++ applications/frontend/reader/src/app/series/[slug]/chapters/(detail)/layout.tsx
@@ -1,21 +1,23 @@
 import { Suspense } from "react";
 import { slugSchema } from "@shared/domains/common/slug";
 import { findBySlug } from "@/actions/series";
-import { findChaptersByIdentifiers } from "@/actions/chapter";
+import { findPublishedChaptersByIdentifiers } from "@/actions/chapter";
 import { ChapterToc } from "@shared/components/organisms/series/chapter/toc";
 import { ChapterLayoutTemplate } from "@shared/components/templates/series/chapter/layout";
 import { ArticleSidebarSkeleton } from "@shared/components/molecules/skeleton";
 
+export const revalidate = 3600;
+
 type Props = {
-  params: Promise<{ slug: string; chapter: string }>;
+  params: Promise<{ slug: string }>;
   children: React.ReactNode;
 };
 
 type TocProps = {
-  params: Promise<{ slug: string; chapter: string }>;
+  params: Promise<{ slug: string }>;
 };
 
-async function AdminChapterTocSection(props: TocProps) {
+async function ChapterTocSection(props: TocProps) {
   const params = await props.params;
   const series = await findBySlug(params.slug);
   const slug = slugSchema.parse(params.slug);
@@ -25,17 +27,17 @@ async function AdminChapterTocSection(props: TocProps) {
       slug={slug}
       seriesTitle={series.title}
       seriesChapterIdentifiers={series.chapters}
-      findChaptersByIdentifiers={findChaptersByIdentifiers}
+      findChaptersByIdentifiers={findPublishedChaptersByIdentifiers}
     />
   );
 }
 
-export default function AdminChapterLayout(props: Props) {
+export default function ChapterLayout(props: Props) {
   return (
     <ChapterLayoutTemplate
       toc={
         <Suspense fallback={<ArticleSidebarSkeleton />}>
-          <AdminChapterTocSection params={props.params} />
+          <ChapterTocSection params={props.params} />
         </Suspense>
       }
     >
